% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.wlasso.R
\name{plot.wlasso}
\alias{plot.wlasso}
\title{Plot weighted LASSO object}
\usage{
\method{plot}{wlasso}(x)
}
\arguments{
\item{x}{object of class "wlasso"}
}
\value{
a graph
}
\description{
Plot weighted LASSO object
}
\examples{
data(simdata_lasso_binomial)
mcv <- wlasso(data = simdata_lasso_binomial, col.y = "y", col.x = 1:50,
              family = "binomial", cluster = "cluster", strata = "strata",
              weights = "weights", method = "dCV", k=10, R=20, seed = 100)
plot(mcv)
}
