\name{WLasso-package}
\alias{WLasso-package}
\alias{WLasso}
\docType{package}
\title{
\packageTitle{WLasso}
}
\description{
\packageDescription{WLasso}
}
\details{
The DESCRIPTION file:
\packageDESCRIPTION{WLasso}
\packageIndices{WLasso}
This package consists of four functions: "Sigma_Estimation.R", "top.R",
"top_thresh.R" and "Whitening_Lasso.R".
For further information on how to use these functions, 
we refer the reader to the vignette of the package.
}
\author{
\packageAuthor{WLasso}

Maintainer: \packageMaintainer{WLasso}
}
\references{
W. Zhu, C. Levy-Leduc, N. Ternes. "A variable selection approach for highly correlated predictors in 
high-dimensional genomic data". arXiv:2007.10768.
}


