% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wlasso.plot.R
\name{wlasso.plot}
\alias{wlasso.plot}
\title{Plot weighted LASSO object}
\usage{
wlasso.plot(x)
}
\arguments{
\item{x}{an object of class "wlasso".}
}
\value{
a graph
}
\description{
Plot weighted LASSO object
}
\examples{
data(simdata_lasso_binomial)
mcv <- wlasso(data = simdata_lasso_binomial, col.y = "y", col.x = 1:50,
              family = "binomial", cluster = "cluster", strata = "strata",
              weights = "weights", method = "dCV", k=10, R=20)
wlasso.plot(mcv)
}
