\name{X}
\alias{X}
\docType{data}
\title{
Example of a design matrix of a linear model}
\description{
It contains an example of a design matrix of a linear model.
}
\usage{data("X")}
\format{
  The format is:
 num [1:50, 1:200] 0.855 1.32 1.018 -0.489 0.234 ...
}
\examples{
data(X)
}
\keyword{datasets}
