\name{Y}
\alias{Y}
\docType{data}
\title{
Example of a response variable of a linear model.
}
\description{
It contains an example of a response variable of a linear model.
}
\usage{data("Y")}
\format{
  The format is:
 num [1:50, 1] 6.027 0.656 3.305 -1.922 -4.943 ...
}
\examples{
data(Y)
}
\keyword{datasets}
