% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wlasso.R
\name{wlasso}
\alias{wlasso}
\title{Weighted LASSO prediction models for complex survey data}
\usage{
wlasso(
  data = NULL,
  col.y = NULL,
  col.x = NULL,
  cluster = NULL,
  strata = NULL,
  weights = NULL,
  design = NULL,
  family = c("gaussian", "binomial"),
  lambda.grid = NULL,
  method = c("dCV", "JKn", "bootstrap", "subbootstrap", "BRR", "split", "extrapolation"),
  k = 10,
  R = 1,
  B = 200,
  cv.error.ind = FALSE,
  train.prob = 0.7,
  method.split = c("dCV", "bootstrap", "subbootstrap"),
  seed = NULL
)
}
\arguments{
\item{data}{A data frame with information about the response variable and covariates, sampling weights and strata and cluster indicators.}

\item{col.y}{column of the response variable. It could be indicated as the column number or the column name.}

\item{col.x}{vector of columns indicating the covariates. It could be defined by numbers or names of the columns.}

\item{cluster}{name of the column indicating clusters}

\item{strata}{name of the column indicating strata}

\item{weights}{name of the column indicating sampling weights}

\item{design}{an object of class \code{survey.design}. It could be \code{NULL} if information about \code{cluster}, \code{strata}, \code{weights} and \code{data} are given.}

\item{family}{family to fit LASSO models, choose between \code{gaussian} or \code{binomial}.}

\item{lambda.grid}{a grid for penalization parameters. If it is not defined by the user, the function will define one.}

\item{method}{method to be applied to define replicate weights, to choose between one of these: \code{JKn}, \code{dCV}, \code{bootstrap}, \code{subbootstrap}, \code{BRR}, \code{split}, \code{extrapolation}}

\item{k}{number of folds to be defined (only for \code{cv}). Default is \code{k=10}.}

\item{R}{number of times the sample is partioned (needed and used only for \code{cv}, \code{split} or \code{extrapolation} methods). Default R=1.}

\item{B}{number of bootstrap resamples (only for \code{bootstrap} and \code{subbootstrap} methods). Default \code{B=200}.}

\item{cv.error.ind}{method for estimating the error for \code{cv} method. FALSE (default) estimates the error for each test set and defines the cross-validated error as the average of all those errors. Option TRUE estimates the cross-validated error as the weighted average of the loss for each unit}

\item{train.prob}{probability for defining training sets (only for \code{split} and \code{extrapolation} methods)}

\item{method.split}{\code{cv} or \code{bootstrap} (only for \code{split} method)}

\item{seed}{define a seed}
}
\value{
An object of class \code{wlasso}.
}
\description{
This function allows as to fit LASSO prediction (linear or logistic) models to complex survey data, considering sampling weights in the estimation process and selects the lambda that minimizes the error based on different replicating weights methods.
}
\examples{
data(simdata_lasso_binomial)
mcv <- wlasso(data = simdata_lasso_binomial,
              col.y = "y", col.x = 1:50,
              family = "binomial",
              cluster = "cluster", strata = "strata", weights = "weights",
              method = "dCV", k=10, R=20, seed = 100)

# Or equivalently:
mydesign <- survey::svydesign(ids=~1, strata = ~strata, weights = ~weights,
                              nest = TRUE, data = simdata_lasso_binomial)
mcv <- wlasso(col.y = "y", col.x = 1:50, design = mydesign,
              family = "binomial",
              method = "dCV", k=10, R=20, seed = 100)
}
